chore: move AddWhatCondModal to legacy
